+++ core/core_blocks.go
@@ -179,81 +179,6 @@ func Identity() Spec {
 	}
 }
 
-// Head emits the first element of the inbound array on one output,
-//and the tail of the array on the other.
-func Head() Spec {
-	return Spec{
-		Name:    "head",
-		Inputs:  []Pin{Pin{"in", ARRAY}},
-		Outputs: []Pin{Pin{"head", ANY}, Pin{"tail", ARRAY}},
-		Kernel: func(in, out, internal MessageMap, s Source, i chan Interrupt) Interrupt {
-			arr, ok := in[0].([]interface{})
-			if !ok {
-				out[0] = NewError("head requires an array")
-				return nil
-			}
-			out[0] = arr[0]
-			out[1] = arr[1:len(arr)]
-			return nil
-		},
-	}
-}
-
-func Len() Spec {
-	return Spec{
-		Name:    "len",
-		Inputs:  []Pin{Pin{"in", ARRAY}},
-		Outputs: []Pin{Pin{"out", NUMBER}},
-		Kernel: func(in, out, internal MessageMap, s Source, i chan Interrupt) Interrupt {
-			arr, ok := in[0].([]interface{})
-			if !ok {
-				out[0] = NewError("len requires an array")
-				return nil
-			}
-			out[0] = float64(len(arr))
-			return nil
-		},
-	}
-}
-
-// Tail emits the last element of the inbound array on one output,
-//and the head of the array on the other.
-func Tail() Spec {
-	return Spec{
-		Name:    "tail",
-		Inputs:  []Pin{Pin{"in", ARRAY}},
-		Outputs: []Pin{Pin{"tail", ANY}, Pin{"head", ARRAY}},
-		Kernel: func(in, out, internal MessageMap, s Source, i chan Interrupt) Interrupt {
-			arr, ok := in[0].([]interface{})
-			if !ok {
-				out[0] = NewError("tail requires an array")
-				return nil
-			}
-			out[0] = arr[len(arr)]
-			out[1] = arr[0 : len(arr)-1]
-			return nil
-		},
-	}
-}
-
-// Append appends the supplied element to the supplied array
-func Append() Spec {
-	return Spec{
-		Name:    "append",
-		Inputs:  []Pin{Pin{"element", ANY}, Pin{"array", ARRAY}},
-		Outputs: []Pin{Pin{"array", ARRAY}},
-		Kernel: func(in, out, internal MessageMap, s Source, i chan Interrupt) Interrupt {
-			arr, ok := in[1].([]interface{})
-			if !ok {
-				out[0] = NewError("Append requires an array")
-				return nil
-			}
-			out[0] = append(arr, in[0])
-			return nil
-		},
-	}
-}
-
 // Merge recursively merges two objects, favouring the first input to resolve conflicts
 func Merge() Spec {
 	return Spec{
